refactor: type 별도 관리 및 type 변경
