+++ README.md
@@ -73,16 +73,17 @@
 
 ## 피드백 바탕 리팩토링 사항
 
-- [ ] 설계 구조 변경
+- [x] 설계 구조 변경
   - [x] 에러 처리 로직
     - [x] 에러 관련 처리 훅으로 분리
 - [x] previewContainer 제거
 - [x] previewCard 내부 로직 수정 (cardType 내부에서 관리)
 - [x] cspell 익스텐션 이용 알파벳 수정
-- [ ] string 배열 타입 -> 튜플 타입 변경
+- [x] string 배열 타입 -> 튜플 타입 변경
+- [x] 타입 별도 파일 관리
 - [x] input props 네이밍 표준 인터페이스 명으로 변경
-- [ ] CardInformation 도메인 -> use- 붙이기 (훅사용)
-- [ ] keyof CardInformationType 수정 (formContainer 파일)
+- [x] CardInformation 도메인 -> use- 붙이기 (훅사용)
+- [x] keyof CardInformationType 수정 (formContainer 파일)
 - [ ] storybook autodocs 주석처리
 - [ ] storybook 타입 에러 (args:any)
 - [x] previewCard uniqueNumber 뒷 8자리 마스킹 처리
