+++ src/components/ErrorMessage/ErrorMessage.tsx
@@ -1,13 +1,9 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
 import Text from "../Text/Text";
+import { errorMessageProps } from "../../types/componentPropsType";
 
-type errorMessageType = {
-  error: boolean;
-  message: string;
-};
-
-const ErrorMessage = ({ error, message }: errorMessageType) => {
+const ErrorMessage = ({ error, message }: errorMessageProps) => {
   return (
     <div css={errorTextWrapperStyle(error)}>
       <Text type="error" text={message || "유효하지 않은 값입니다."} />
