+++ src/components/FormContainer/FormContainer.tsx
@@ -1,48 +1,15 @@
 /** @jsxImportSource @emotion/react */
 import FormSection from "../FormSection/FormSection";
 import { css } from "@emotion/react";
-import { CardInformationType, setCardInformationType } from "../../types/CardInformationType";
-
-type FormContainerProps = {
-  cardInformationState: CardInformationType;
-  setCardInformationState: setCardInformationType;
-};
-
-const UIControllerData = {
-  uniqueNumber: {
-    title: "결제할 카드 번호를 입력해 주세요",
-    description: "본인 명의의 카드만 결제 가능합니다.",
-    inputFieldData: {
-      label: "카드번호",
-      inputNumber: 4,
-      inputProps: { placeholder: ["1234", "1234", "1234", "1234"], maxLength: 4 },
-    },
-  },
-  expirationDate: {
-    title: "카드 유효기간을 입력해 주세요",
-    description: "월/년도(MMYY)를 순서대로 입력해 주세요.",
-    inputFieldData: {
-      label: "유효기간",
-      inputNumber: 2,
-      inputProps: { placeholder: ["MM", "YY"], maxLength: 2 },
-    },
-  },
-  cvcNumber: {
-    title: "CVC 번호를 입력해 주세요",
-    description: "",
-    inputFieldData: {
-      label: "cvc",
-      inputNumber: 1,
-      inputProps: { placeholder: ["123"], maxLength: 3 },
-    },
-  },
-};
+import { FormContainerProps } from "../../types/componentPropsType";
+import formUIControllerData from "../../constants/formUIControllerData";
+import { CardInformationType } from "../../types/CardInformationType";
 
 const FormContainer = ({ cardInformationState, setCardInformationState }: FormContainerProps) => {
   return (
     <div css={FormContainerStyle}>
-      {(Object.keys(cardInformationState) as (keyof typeof cardInformationState)[]).map((key) => {
-        const formSectionData = UIControllerData[key];
+      {(Object.keys(cardInformationState) as (keyof CardInformationType)[]).map((key) => {
+        const formSectionData = formUIControllerData[key];
         return (
           <FormSection
             title={formSectionData.title}
