+++ src/types/CardInformationType.ts
@@ -1,7 +1,7 @@
 export type CardInformationType = {
-  uniqueNumber: string[];
-  expirationDate: string[];
-  cvcNumber: string[];
+  uniqueNumber: [string, string, string, string];
+  expirationDate: [string, string];
+  cvcNumber: [string];
 };
 
 export type setCardInformationType = React.Dispatch<React.SetStateAction<CardInformationType>>;
